adding basic class
